Manage the application database user, setting privileges and hosts
The default application database user will be created with only
user-level mysql privileges, and by default is only permitted to
connect from localhost.

Priviliges and allowed hosts can be configured with attributes,
and you can add additional resources in application cookbooks if
grants are required on other databases or tables.